AboutPrimitives fixing line 92

+++ koans/src/beginner/AboutPrimitives.java
@@ -89,8 +89,8 @@ public class AboutPrimitives {
 
     @Koan
     public void bytesHaveASmallerRangeThanShorts() {
-        assertEquals(Byte.MIN_VALUE, -128);
-        assertEquals(Byte.MAX_VALUE, 127);
+        assertEquals(Byte.MIN_VALUE, (byte)-128);
+        assertEquals(Byte.MAX_VALUE, (byte)127);
 
         // Why would you use short or byte considering that you need to do explicit casts?
     }
@@ -189,7 +189,7 @@ public class AboutPrimitives {
 
     @Koan
     public void floatSize() {
-        assertEquals(Float.SIZE, 32);
+        assertEquals(Float.SIZE, 32f);
     }
 
     private Class<?> getType(int value) {
